requirements.txt
webdataset
pytorch-lightning
tensorboard
polyscope
